requirements.txt
django
django-froala-editor